requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==7.1.0
docutils==0.19
ffpyplayer==4.3.5
frozenlist==1.3.3
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
kivyunittest==0.1.8
multidict==6.0.4
Pillow==9.4.0
Pygments==2.14.0
python-reaperdaw==0.0.7
requests==2.28.2
urllib3==1.26.14
yarl==1.8.2
